% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.r
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{number of people in the room}
}
\value{
The probability that two or more people in a class of size x have the same birthday.
}
\description{
Birthday
}
\examples{
birthday(20:25)
}
