% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{x1000 dollars in damage}
}
}
\source{
From textbook
}
\usage{
fire
}
\description{
A dataset collected concerning home damage from fire
}
\keyword{datasets}
