% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of bootstrap replications}

\item{x}{sample data}

\item{fun}{statistic to compute}

\item{alpha}{significance level (1 - alpha is ci)}

\item{cx}{the plot's text size}

\item{...}{optional plotting parameters}
}
\value{
a list with ci, the function used, and the original data
}
\description{
myboot2
}
\examples{
myboot2(x=ddt$DDT)
}
