% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter = 10000, a = 0, b = 10)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{number of iterations}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
a histogram of the mean sample along with a density curve and theoretical normal curve
}
\description{
mycltu
}
\examples{
mycltu(n=20,iter=100000)
}
