% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the function}

\item{sigma}{the standard deviation of the function}

\item{a}{the x value to compute the probability}
}
\value{
the probability P(X≤a) and the graph
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
