% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overbooking.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{probability of overbooking - the "pain" factor}

\item{p}{probability a passenger shows up}
}
\value{
list containing optimal number of seats sold via discrete and
continuous models, plots the two models as well
}
\description{
Overbooking
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
